requirements.txt
# KodaCare - Backend Dependencies
flask>=3.0.0
flask-cors>=4.0.0
flask-jwt-extended>=4.6.0
pymongo>=4.6.0
python-dotenv>=1.0.0
google-genai>=1.0.0
bcrypt>=4.1.0
Werkzeug>=3.0.0
